split i18n translations into one file per language
